+++ database/migrations/2016_11_02_180355_create_sections_table.php
@@ -15,7 +15,7 @@ class CreateSectionsTable extends Migration
     {
         Schema::create('sections', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('book_id');
+            $table->bigInteger('book_id');
             $table->string('name');
             $table->bigInteger('number');
             $table->timestamps();
